+++ tests/docker/runners/bash/cluster-tls.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD" "REDIS_SENTINEL_PASSWORD", "FRED_TEST_TLS_CREDS")
+declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD", "FRED_TEST_TLS_CREDS")
 
 for env in "${arr[@]}"
 do
@@ -10,20 +10,12 @@ do
   fi
 done
 
-FEATURES="enable-native-tls ignore-auth-error"
-
-export FRED_REDIS_CENTRALIZED_HOST="example.com"
-export FRED_REDIS_CENTRALIZED_PORT="6379"
-export FRED_REDIS_CLUSTER_HOST="node-30001.example.com"
-export FRED_REDIS_CLUSTER_PORT="30001"
-export FRED_TEST_TLS_CREDS=$PWD/tests/tmp/creds
-export FRED_CI_TLS="true"
-
+FEATURES="enable-native-tls vendored-openssl ignore-auth-error"
 # https://github.com/sfackler/rust-native-tls/issues/143
 echo "This may not work on Mac"
 
 if [ -z "$FRED_CI_NEXTEST" ]; then
-  cargo test --release --lib --tests --features "$FEATURES" -- --test-threads=1 "$@"
+  FRED_CI_TLS=true cargo test --release --lib --tests --features "$FEATURES" -- --test-threads=1 "$@"
 else
-  cargo nextest run --release --lib --tests --features "$FEATURES" --test-threads=1 "$@"
+  FRED_CI_TLS=true cargo nextest run --release --lib --tests --features "$FEATURES" --test-threads=1 "$@"
 fi