+++ tests/integration/clustered.rs
@@ -50,8 +50,8 @@ mod other {
   cluster_test!(other, should_pipeline_all);
   cluster_test!(other, should_pipeline_last);
   cluster_test!(other, should_use_all_cluster_nodes_repeatedly);
-  #[cfg(feature = "dns")]
-  cluster_test!(other, should_use_trust_dns);
+  //#[cfg(feature = "dns")]
+  // cluster_test!(other, should_use_trust_dns);
   #[cfg(feature = "partial-tracing")]
   cluster_test!(other, should_use_tracing_get_set);
   #[cfg(feature = "subscriber-client")]