+++ tests/runners/all-features.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-docker-compose -f tests/docker/compose/centralized.yml \
+TEST_ARGV="$1" docker-compose -f tests/docker/compose/centralized.yml \
   -f tests/docker/compose/cluster.yml \
-  -f tests/docker/runners/compose/all-features.yml run --rm all-features-tests
+  -f tests/docker/runners/compose/all-features.yml run -u $(id -u ${USER}):$(id -g ${USER}) --rm all-features-tests