+++ tests/runners/sentinel-features.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 
-docker-compose -f tests/docker/compose/sentinel.yml -f tests/docker/runners/compose/sentinel-features.yml run --rm sentinel-tests
+TEST_ARGV="$1" docker-compose -f tests/docker/compose/sentinel.yml -f tests/docker/runners/compose/sentinel-features.yml \
+  run -u $(id -u ${USER}):$(id -g ${USER}) --rm sentinel-tests